// Frontend/src/app/erp/models/servicesDetails.model.ts
export class ServicesDetails{
    idDetallePedido: number;
      linea: number;
      idMaterial: number;
      nombre: string;
      precio: number;
      color?: string;
}

export class CServicesDetails{
      linea: number;
      idMaterial: number;
      nombre: string;
      precio: number;
      color?: string;
}